+++ js/create-advert.js
@@ -1,3 +1,7 @@
+import {getLocationPoint} from './create-location.js';
+import {getAuthor} from './create-user.js';
+import {getOffer} from './create-offer.js';
+
 const getAdvert = () => {
   const locationPoint = getLocationPoint();
   return {
