+++ js/create-links.js
@@ -1,3 +1,5 @@
+import {createRandomIdFromRangeGenerator} from './util.js';
+import {AMOUNT_USERS} from './data.js';
 
 const createRandomLinksId = createRandomIdFromRangeGenerator(1, AMOUNT_USERS);
 
@@ -5,4 +7,5 @@ const getRandomLinksId = () => {
   const currentLinksId = createRandomLinksId();
   return (currentLinksId < AMOUNT_USERS) ? `${0}${currentLinksId}` : currentLinksId;
 };
+
 export {getRandomLinksId};
