+++ js/create-location.js
@@ -1,3 +1,5 @@
+import {getRandomPoint} from './create-coordinates.js';
+
 const getLocationPoint = () => ({
   lat: getRandomPoint(35.65000, 35.70000, 5),
   lng: getRandomPoint(139.70000, 139.80000, 5),
