+++ js/create-user.js
@@ -1,3 +1,5 @@
+import {getRandomLinksId} from './create-links.js';
+
 const getAuthor = () => ({
   avatar: `img/avatars/user${getRandomLinksId()}.png`
 });
