+++ js/main.js
@@ -1,0 +1,1 @@
+import './create-adverts-list.js';
