+++ package.json
@@ -16,6 +16,7 @@
     "eslint-config-htmlacademy": "9.0.0"
   },
   "license": "MIT",
+  "type": "module",
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
